requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
slowapi